Purchase in DB, GroupPreview in MainActivity

+++ app/src/main/java/group3/myapplicationlab2/Purchase.java
@@ -19,7 +19,7 @@ public class Purchase {
     private String path_image = "nopath";
     private Long date_millis;
     private String causal;
-    private Float total_amount;
+    private double total_amount;
     private String author_name;
     private String group_id;
     private String date;
@@ -43,8 +43,8 @@ public class Purchase {
     public String getAuthorName(){return this.author_name;}
     public void setAuthorName(String an){this.author_name = an;}
 
-    public float getTotalAmount() {return this.total_amount;}
-    public void setTotalAmount(float ta){this.total_amount = ta;}
+    public double getTotalAmount() {return this.total_amount;}
+    public void setTotalAmount(double ta){this.total_amount = ta;}
 
     public String getCausal() {return this.causal;}
     public void setCausal(String causal){this.causal = causal;}
